started refactoring boilerplates
