+++ client/components/EntryList.js
