+++ client/components/Thermometer.js
