+++ client/config/auth.js
