+++ client/pages/Journal.js
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+const Journal = () => {
+  return <p>Journal Page</p>;
+};
+export default Journal;
