+++ client/pages/Meditation.js
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+const Meditation = () => {
+  return <p>Meditation Page</p>;
+};
+
+export default Meditation;
