+++ client/pages/Welcome.js
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+const Welcome = () => {
+  return <p>Welcome Page</p>;
+};
+export default Welcome;
