+++ server/controllers/entryController.js
