+++ sql/models/Entry.js
